server/requirements.txt
fastapi
uvicorn[standard]
python-multipart
torch
torchvision
ultralytics
git+https://github.com/openai/CLIP.git
faiss-cpu
opencv-python
pillow
numpy
